完成图书存入数据库

+++ aituidao_server/models.py
@@ -9,5 +9,5 @@ class Book(models.Model):
     cover = models.CharField(max_length = 100, blank=True, null=True)
     filename = models.CharField(max_length = 100)
     pushCount = models.IntegerField()
-    doubanRate = models.FloatField()
+    doubanRate = models.IntegerField()
     deleted = models.BooleanField()
